Generate a meaningful source_ref for documentation links
Rather than hard-code the source code URLs to point to the master
branch, we should instead reference the exact Git commit that reflects
the current state of the repository. That way, links will not be broken
even if we were to build an older version of the documentation.

If an existing tag happens to point to the current commit object, the
`source_ref/0` function will instead reference the tag's name rather
than full the SHA-1 hash.