Add a border of deactivated squares
The purpose of the border is so our code can always look at the adjacent
squares without having to check if we're at the edge of the board.

I'm already seeing how my representation of the board is going to become
problematic as it involves the creation and passing around of the entire
structure more than it should and will likely be difficult to implement
further functions for the game. I might have to switch to processes to
store the square states, but will think about this some more.