Use a map with coordinate tuple keys to represent the board
The list approach was more difficult than it should be and not very
efficient. I believe using a map with `{row, col}` keys pointing to each
`%Square{}` should make things much easier.

I've also created a `%Board{}` struct to hold the squares and additional
metadata. This also means that I'll be able to convert the `debug/2`
function into an implementation of the _Inspect_ protocol (currently the
debug output is broken, but I wanted to commit this data structure
change).

An Agent is now used to store the board state and will provide a good
abstraction layer to build the public API upon.