Force a minimal word list when running tests
The default word list is very large and takes quite a bit of time to
work with, which slows down testing substantially. I'm not sure if this
is the best way to go about this, but testing via `mix test --no-start`
won't work once things start getting a bit more complex, so I need to do
something. Might have to revisit this change later.